Add files via upload
This project involves fine-tuning the LLaMA 3 model, which is an advanced language model. The main objective of this project is to improve the model's performance and adapt it to specific data and requirements. This process allows us to optimize the model for particular tasks and increase its accuracy and efficiency.

A crucial aspect of this project is the use of JSON format for data.
Furthermore, specialized LLaMA Factory libraries are also used:

llamafactory.chat: This module is used for utilizing chat models. It's likely used to create text-based interactions with the fine-tuned model.
llamafactory.extras.misc: This module is used for memory management using torch_gc. This helps optimize resource consumption and efficiently manage memory throughout the fine-tuning process.